requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
torch==2.0.1
torchvision==0.15.2
diffusers==0.21.4
transformers==4.34.0
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
werkzeug==2.3.7
flask-cors==4.0.0
accelerate==0.23.0
xformers==0.0.22
flask-jwt-extended==4.5.3
redis==5.0.0